generators/image/requirements.txt
certifi==2018.11.29
chardet==3.0.4
Django==1.11.20
idna==2.8
image==1.5.27
numpy==1.16.2
Pillow==5.4.1
pytz==2018.9
requests==2.21.0
urllib3==1.24.1
